requirements.txt
altgraph==0.16.1

appdirs==1.4.3
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.4.1
crayons==0.2.0
Django==1.10.3
docopt==0.6.2
enum34==1.1.6
Flask==0.12.2
Flask-SSLify==0.1.5
future==0.17.1
idna==2.6
image==1.5.27
itsdangerous==0.24
Jinja2==2.10
macholib==1.11
mysql-connector==2.2.9
numpy==1.19.5
opencv-python==4.3.0.38
pefile==2019.4.18
piexif==1.1.2
Pillow==6.0.0
pyee==6.0.0
PyInstaller==3.4
pymongo==3.6.1
pyppeteer==0.0.25
pyTelegramBotAPI==3.6.3
pywin32-ctypes==0.2.0
requests==2.18.4
schedule==0.5.0
selenium==3.141.0
six==1.11.0
soupsieve==1.9.1
spark-parser==1.8.7
Stegano==0.9.2
telebot==0.0.3
tqdm==4.31.1
uncompyle6==3.3.2
urllib3==1.22
vk-api==11.6.0
websockets==7.0
Werkzeug==0.14.1
xdis==4.0.1
